+++ src/triggers/sharepoint-folder/index.ts
@@ -1,5 +1,8 @@
 export {
   SharePointFolderTrigger,
   SHAREPOINT_FOLDER_TRIGGER_TYPE,
+  type FileReceivedEvent,
+  type FileReceivedResult,
+  type OnFileReceivedFn,
 } from './sharepoint-folder-trigger';
 export * from './graph-client';